refine: getprocinfo

+++ interfaces/innerkits/unwinder/include/dfx_map.h
@@ -28,7 +28,7 @@ class DfxElf;
 
 class DfxMap {
 public:
-    static std::shared_ptr<DfxMap> Create(const std::string buf, int size);
+    static std::shared_ptr<DfxMap> Create(const std::string buf, size_t size);
     static void PermsToProts(const std::string perms, uint32_t& prots, uint32_t& flag);
 
     DfxMap() = default;
@@ -39,7 +39,7 @@ public:
         uint32_t prots, const std::string& name)
         : begin(begin), end(end), offset(offset), prots(prots), name(name) {}
 
-    bool Parse(const std::string buf, int size);
+    bool Parse(const std::string buf, size_t size);
     bool IsMapExec();
     bool IsArkExecutable();
     const std::shared_ptr<DfxElf> GetElf(pid_t pid = 0);
